requirements.txt
livekit
livekit-agents
flask
flask-cors
werkzeug
asyncio
aiohttp
python-dotenv
livekit-plugins-google
